+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **Bug Fixes:**
 
 - When `validate_checksum` is disabled (default), an infinite loop and memory leak when the data_size of an event is set to zero was fixed. (omerbenamram/evtx#264)
+- `-t, --threads` was not working in the `computer-metrics` and `search` commands. (#1563) (@hach1yon)
 
 ## 3.6.0 [2025/09/25] - Nezamezuki Release